backend/requirement.txt
flask
flask-cors
pdfplumber
PyMuPDF    # For 'import fitz'
Pillow     # For 'from PIL import Image'
pytesseract
langchain-core>=1.0.0
langchain-community>=0.2.0
langchain-text-splitters>=1.0.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0
chromadb
python-dotenv  # For 'load_dotenv()'
colorama       # For colored logs